Refactoring the MixView class; trying to reduce code and organize it bit better
