+++ src/org/mixare/MixView.java
@@ -18,13 +18,7 @@
  */
 package org.mixare;
 
-/**
- * This class is the main application which uses the other classes for different
- * functionalities.
- * It sets up the camera screen and the augmented screen which is in front of the
- * camera screen.
- * It also handles the main sensor events, touch events and location events.
- */
+import org.mixare.utils.ErrorUtility;
 
 import static android.hardware.SensorManager.SENSOR_DELAY_GAME;
 
@@ -62,7 +56,6 @@ import android.os.PowerManager;
 import android.os.PowerManager.WakeLock;
 import android.provider.Settings;
 import android.util.Log;
-import android.view.Display;
 import android.view.Gravity;
 import android.view.KeyEvent;
 import android.view.Menu;
@@ -72,7 +65,6 @@ import android.view.SurfaceHolder;
 import android.view.SurfaceView;
 import android.view.View;
 import android.view.Window;
-import android.view.WindowManager;
 import android.view.View.OnTouchListener;
 import android.view.ViewGroup.LayoutParams;
 import android.widget.FrameLayout;
@@ -80,7 +72,15 @@ import android.widget.SeekBar;
 import android.widget.TextView;
 import android.widget.Toast;
 
-public class MixView extends Activity implements SensorEventListener, OnTouchListener{
+/**
+ * This class is the main application which uses the other classes for different
+ * functionalities.
+ * It sets up the camera screen and the augmented screen which is in front of the
+ * camera screen.
+ * It also handles the main sensor events, touch events and location events.
+ */
+public class MixView extends Activity
+                     implements SensorEventListener, OnTouchListener {
 
 	private CameraSurface camScreen;
 	private AugmentedView augScreen;
@@ -103,7 +103,6 @@ public class MixView extends Activity implements SensorEventListener, OnTouchLis
 
 	private Matrix tempR = new Matrix();
 	private Matrix finalR = new Matrix();
-	private Matrix histR[] = new Matrix[60];
 	private Matrix m1 = new Matrix();
 	private Matrix m2 = new Matrix();
 	private Matrix m3 = new Matrix();
@@ -111,8 +110,6 @@ public class MixView extends Activity implements SensorEventListener, OnTouchLis
 
 	private SeekBar myZoomBar;
 	private WakeLock mWakeLock;
-
-	private boolean fError;
 
 	private int compassErrorDisplayed = 0;
 
@@ -143,70 +140,16 @@ public class MixView extends Activity implements SensorEventListener, OnTouchLis
 		return zoomProgress;
 	}
 
-	public void doError(Exception ex1) {
-		if (!fError) {
-			fError = true;
 
-			setErrorDialog();
-
-			ex1.printStackTrace();
-			try {
-			} catch (Exception ex2) {
-				ex2.printStackTrace();
-			}
-		}
-
-		try {
-			augScreen.invalidate();
-		} catch (Exception ignore) {
-		}
-	}
-
-	public void killOnError() throws Exception {
-		if (fError)
-			throw new Exception();
-	}
+//	public void killOnError() throws Exception {
+//		if (fError)
+//			throw new Exception();
+//	}
 
 	public void repaint() {
 		dataView = new DataView(mixContext);
 		dWindow = new PaintScreen();
 		setZoomLevel();
-	}
-
-	public void setErrorDialog(){
-		AlertDialog.Builder builder = new AlertDialog.Builder(this);
-		builder.setMessage(getString(DataView.CONNECTION_ERROR_DIALOG_TEXT));
-		builder.setCancelable(false);
-
-		/*Retry*/
-		builder.setPositiveButton(DataView.CONNECTION_ERROR_DIALOG_BUTTON1, new DialogInterface.OnClickListener() {
-			public void onClick(DialogInterface dialog, int id) {
-				fError=false;
-				//TODO improve
-				try {
-					repaint();	       		
-				}
-				catch(Exception ex){
-					//Don't call doError, it will be a recursive call.
-					//doError(ex);
-				}
-			}
-		});
-		/*Open settings*/
-		builder.setNeutralButton(DataView.CONNECTION_ERROR_DIALOG_BUTTON2, new DialogInterface.OnClickListener() {
-			public void onClick(DialogInterface dialog, int id) {
-				Intent intent1 = new Intent(Settings.ACTION_WIRELESS_SETTINGS); 
-				startActivityForResult(intent1, 42);
-			}
-		});
-		/*Close application*/
-		builder.setNegativeButton(DataView.CONNECTION_ERROR_DIALOG_BUTTON3, new DialogInterface.OnClickListener() {
-			public void onClick(DialogInterface dialog, int id) {
-				System.exit(0);
-			}
-		});
-		AlertDialog alert = builder.create();
-		alert.show();
 	}
 
 	@Override
@@ -225,7 +168,7 @@ public class MixView extends Activity implements SensorEventListener, OnTouchLis
 			this.mWakeLock = pm.newWakeLock(
 					PowerManager.SCREEN_BRIGHT_WAKE_LOCK, "My Tag");
 
-			killOnError();
+			//killOnError();
 			requestWindowFeature(Window.FEATURE_NO_TITLE);
 
 			/*Get the preference file PREFS_NAME stored in the internal memory of the phone*/
@@ -298,8 +241,9 @@ public class MixView extends Activity implements SensorEventListener, OnTouchLis
 
 			} 
 
-		} catch (Exception ex) {
-			doError(ex);
+		} catch (Exception e) {
+			ErrorUtility.handleError(TAG, e, true);
+      augScreen.invalidate();
 		}
 	}
 
@@ -351,11 +295,11 @@ public class MixView extends Activity implements SensorEventListener, OnTouchLis
 
       unregisterListners();
 
-			if (fError) {
-				finish();
-			}
-		} catch (Exception ex) {
-			doError(ex);
+//			if (fError) {
+//				finish();
+//			}
+		} catch (Exception e) {
+			ErrorUtility.handleError(TAG, e, false);
 		}
 	}
 
@@ -366,7 +310,7 @@ public class MixView extends Activity implements SensorEventListener, OnTouchLis
 		try {
 			this.mWakeLock.acquire();
 
-			killOnError();
+			//killOnError();
 			SharedPreferences settings = getSharedPreferences(PREFS_NAME, 0);
 			mixContext.mixView = this;
 			dataView.doStart();
@@ -408,10 +352,6 @@ public class MixView extends Activity implements SensorEventListener, OnTouchLis
 			
 			m4.toIdentity();
 
-			for (int i = 0; i < histR.length; i++) {
-				histR[i] = new Matrix();
-			}
-
 			sensorMgr = (SensorManager) getSystemService(SENSOR_SERVICE);
 
 			sensors = sensorMgr.getSensorList(Sensor.TYPE_ACCELEROMETER);
@@ -444,8 +384,8 @@ public class MixView extends Activity implements SensorEventListener, OnTouchLis
 			}
 			downloadThread = new Thread(mixContext.downloadManager);
 			downloadThread.start();
-		} catch (Exception ex) {
-			doError(ex);
+		} catch (Exception e) {
+			ErrorUtility.handleError(TAG, e, false);
       unregisterListners();
 		}
 
@@ -685,6 +625,7 @@ public class MixView extends Activity implements SensorEventListener, OnTouchLis
 			}
 
 		} catch (Exception e) {
+			ErrorUtility.handleError(TAG, e, false);
 			Log.e(TAG, e.getStackTrace().toString());
 		}
 	}
@@ -692,18 +633,19 @@ public class MixView extends Activity implements SensorEventListener, OnTouchLis
 	@Override
 	public boolean onTouchEvent(MotionEvent me) {
 		try {
-			killOnError();
+			//killOnError();
 
 			float xPress = me.getX();
 			float yPress = me.getY();
-			if (me.getAction() == MotionEvent.ACTION_UP) {
+
+			if (me.getAction() == MotionEvent.ACTION_UP)
 				dataView.clickEvent(xPress, yPress);
-			}
 
 			return true;
-		} catch (Exception ex) {
-			//doError(ex);
-			ex.printStackTrace();
+
+		} catch (Exception e) {
+			ErrorUtility.handleError(TAG, e, false);
+
 			return super.onTouchEvent(me);
 		}
 	}
@@ -711,7 +653,7 @@ public class MixView extends Activity implements SensorEventListener, OnTouchLis
 	@Override
 	public boolean onKeyDown(int keyCode, KeyEvent event) {
 		try {
-			killOnError();
+			//killOnError();
 
 			if (keyCode == KeyEvent.KEYCODE_BACK) {
 				if (dataView.isDetailsView()) {
@@ -774,220 +716,10 @@ public class MixView extends Activity implements SensorEventListener, OnTouchLis
         if(mixContext.downloadManager != null)
           mixContext.downloadManager.stop();
       }
-    } catch (Exception ignore) {
-      /* Exception igonred */
+
+    } catch (Exception e) {
+      /* Exception ignored, but logged */
+      ErrorUtility.handleError(TAG, e, false);
 		}
   }
-}
-
-/**
- * @author daniele
- *
- */
-class CameraSurface extends SurfaceView implements SurfaceHolder.Callback {
-	MixView app;
-	SurfaceHolder holder;
-	Camera camera;
-
-	CameraSurface(Context context) {
-		super(context);
-
-		try {
-			app = (MixView) context;
-
-			holder = getHolder();
-			holder.addCallback(this);
-			holder.setType(SurfaceHolder.SURFACE_TYPE_PUSH_BUFFERS);
-		} catch (Exception ex) {
-
-		}
-	}
-
-	public void surfaceCreated(SurfaceHolder holder) {
-		try {
-			if (camera != null) {
-				try {
-					camera.stopPreview();
-				} catch (Exception ignore) {
-				}
-				try {
-					camera.release();
-				} catch (Exception ignore) {
-				}
-				camera = null;
-			}
-
-			camera = Camera.open();
-			camera.setPreviewDisplay(holder);
-		} catch (Exception ex) {
-			try {
-				if (camera != null) {
-					try {
-						camera.stopPreview();
-					} catch (Exception ignore) {
-					}
-					try {
-						camera.release();
-					} catch (Exception ignore) {
-					}
-					camera = null;
-				}
-			} catch (Exception ignore) {
-
-			}
-		}
-	}
-
-	public void surfaceDestroyed(SurfaceHolder holder) {
-		try {
-			if (camera != null) {
-				try {
-					camera.stopPreview();
-				} catch (Exception ignore) {
-				}
-				try {
-					camera.release();
-				} catch (Exception ignore) {
-				}
-				camera = null;
-			}
-		} catch (Exception ex) {
-			ex.printStackTrace();
-		}
-	}
-
-
-	public void surfaceChanged(SurfaceHolder holder, int format, int w, int h) {
-		try {
-			Camera.Parameters parameters = camera.getParameters();
-			try {
-				List<Camera.Size> supportedSizes = null;
-				//On older devices (<1.6) the following will fail
-				//the camera will work nevertheless
-				supportedSizes = Compatibility.getSupportedPreviewSizes(parameters);
-
-				//preview form factor
-				float ff = (float)w/h;
-				Log.d("Mixare", "Screen res: w:"+ w + " h:" + h + " aspect ratio:" + ff);
-
-				//holder for the best form factor and size
-				float bff = 0;
-				int bestw = 0;
-				int besth = 0;
-				Iterator<Camera.Size> itr = supportedSizes.iterator();
-
-				//we look for the best preview size, it has to be the closest to the
-				//screen form factor, and be less wide than the screen itself
-				//the latter requirement is because the HTC Hero with update 2.1 will
-				//report camera preview sizes larger than the screen, and it will fail
-				//to initialize the camera
-				//other devices could work with previews larger than the screen though
-				while(itr.hasNext()) {
-					Camera.Size element = itr.next();
-					//current form factor
-					float cff = (float)element.width/element.height;
-					//check if the current element is a candidate to replace the best match so far
-					//current form factor should be closer to the bff
-					//preview width should be less than screen width
-					//preview width should be more than current bestw
-					//this combination will ensure that the highest resolution will win
-					Log.d("Mixare", "Candidate camera element: w:"+ element.width + " h:" + element.height + " aspect ratio:" + cff);
-					if ((ff-cff <= ff-bff) && (element.width <= w) && (element.width >= bestw)) {
-						bff=cff;
-						bestw = element.width;
-						besth = element.height;
-					}
-				} 
-				Log.d("Mixare", "Chosen camera element: w:"+ bestw + " h:" + besth + " aspect ratio:" + bff);
-				//Some Samsung phones will end up with bestw and besth = 0 because their minimum preview size is bigger then the screen size.
-				//In this case, we use the default values: 480x320
-				if ((bestw == 0) || (besth == 0)){
-					Log.d("Mixare", "Using default camera parameters!");
-					bestw = 480;
-					besth = 320;
-				}
-				parameters.setPreviewSize(bestw, besth);
-			} catch (Exception ex) {
-				parameters.setPreviewSize(480 , 320);
-			}
-
-			camera.setParameters(parameters);
-			camera.startPreview();
-		} catch (Exception ex) {
-			ex.printStackTrace();
-		}
-	}
-}
-
-class AugmentedView extends View {
-	MixView app;
-	int xSearch=200;
-	int ySearch = 10;
-	int searchObjWidth = 0;
-	int searchObjHeight=0;
-
-	public AugmentedView(Context context) {
-		super(context);
-
-		try {
-			app = (MixView) context;
-
-			app.killOnError();
-		} catch (Exception ex) {
-			app.doError(ex);
-		}
-	}
-
-	@Override
-	protected void onDraw(Canvas canvas) {
-		try {
-			//			if (app.fError) {
-			//
-			//				Paint errPaint = new Paint();
-			//				errPaint.setColor(Color.RED);
-			//				errPaint.setTextSize(16);
-			//				
-			//				/*Draws the Error code*/
-			//				canvas.drawText("ERROR: ", 10, 20, errPaint);
-			//				canvas.drawText("" + app.fErrorTxt, 10, 40, errPaint);
-			//
-			//				return;
-			//			}
-
-			app.killOnError();
-
-			MixView.dWindow.setWidth(canvas.getWidth());
-			MixView.dWindow.setHeight(canvas.getHeight());
-
-			MixView.dWindow.setCanvas(canvas);
-
-			if (!MixView.dataView.isInited()) {
-				MixView.dataView.init(MixView.dWindow.getWidth(), MixView.dWindow.getHeight());
-			}
-			if (app.isZoombarVisible()){
-				Paint zoomPaint = new Paint();
-				zoomPaint.setColor(Color.WHITE);
-				zoomPaint.setTextSize(14);
-				String startKM, endKM;
-				endKM = "80km";
-				startKM = "0km";
-				/*if(MixListView.getDataSource().equals("Twitter")){
-					startKM = "1km";
-				}*/
-				canvas.drawText(startKM, canvas.getWidth()/100*4, canvas.getHeight()/100*85, zoomPaint);
-				canvas.drawText(endKM, canvas.getWidth()/100*99+25, canvas.getHeight()/100*85, zoomPaint);
-
-				int height= canvas.getHeight()/100*85;
-				int zoomProgress = app.getZoomProgress();
-				if (zoomProgress > 92 || zoomProgress < 6) {
-					height = canvas.getHeight()/100*80;
-				}
-				canvas.drawText(app.getZoomLevel(),  (canvas.getWidth())/100*zoomProgress+20, height, zoomPaint);
-			}
-
-			MixView.dataView.draw(MixView.dWindow);
-		} catch (Exception ex) {
-			app.doError(ex);
-		}
-	}
 }
